site/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
waitress==3.0.0
playwright==1.47.2
pillow==10.4.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.10
numpy==1.26.4
tzdata==2024.2
python-dotenv==1.0.1
discord.py==2.4.0
